+++ test/serializer/generic_test.g.dart
@@ -1,40 +1,54 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of example.serialize_ciclical;
+part of serializer.generic;
 
 // **************************************************************************
 // DsonGenerator
 // **************************************************************************
 
-abstract class _$SomeObjectSerializable<T> extends SerializableMap {
-  int get id;
-  List<T> get genericList;
-  void set id(int v);
-  void set genericList(List<T> v);
+abstract class _$PageSerializable<T> extends SerializableMap {
+  int get size;
+  int get total;
+  int get number;
+  List<T> get items;
+  void set size(int v);
+  void set total(int v);
+  void set number(int v);
+  void set items(List<T> v);
 
   operator [](Object __key) {
     switch (__key) {
-      case 'id':
-        return id;
-      case 'genericList':
-        return genericList;
+      case 'size':
+        return size;
+      case 'total':
+        return total;
+      case 'number':
+        return number;
+      case 'items':
+        return items;
     }
-    throwFieldNotFoundException(__key, 'SomeObject');
+    throwFieldNotFoundException(__key, 'Page');
   }
 
   operator []=(Object __key, __value) {
     switch (__key) {
-      case 'id':
-        id = __value;
+      case 'size':
+        size = __value;
         return;
-      case 'genericList':
-        genericList = fromSerialized(__value, () => new List<T>());
+      case 'total':
+        total = __value;
+        return;
+      case 'number':
+        number = __value;
+        return;
+      case 'items':
+        items = fromSerialized(__value, () => new List<T>());
         return;
     }
-    throwFieldNotFoundException(__key, 'SomeObject');
+    throwFieldNotFoundException(__key, 'Page');
   }
 
-  Iterable<String> get keys => SomeObjectClassMirror.fields.keys;
+  Iterable<String> get keys => PageClassMirror.fields.keys;
 }
 
 abstract class _$PersonSerializable extends SerializableMap {
@@ -72,29 +86,32 @@ abstract class _$PersonSerializable extends SerializableMap {
 // MirrorsGenerator
 // **************************************************************************
 
-_SomeObject__Constructor([positionalParams, namedParams]) => new SomeObject();
+_Page__Constructor([positionalParams, namedParams]) => new Page();
 
-const $$SomeObject_fields_id = const DeclarationMirror(name: 'id', type: int);
-const $$SomeObject_fields_genericList =
-    const DeclarationMirror(name: 'genericList', type: const [List, dynamic]);
+const $$Page_fields_size = const DeclarationMirror(name: 'size', type: int);
+const $$Page_fields_total = const DeclarationMirror(name: 'total', type: int);
+const $$Page_fields_number = const DeclarationMirror(name: 'number', type: int);
+const $$Page_fields_items =
+    const DeclarationMirror(name: 'items', type: const [List, dynamic]);
 
-const SomeObjectClassMirror = const ClassMirror(
-    name: 'SomeObject',
-    constructors: const {
-      '': const FunctionMirror(name: '', $call: _SomeObject__Constructor)
-    },
-    fields: const {
-      'id': $$SomeObject_fields_id,
-      'genericList': $$SomeObject_fields_genericList
-    },
-    getters: const [
-      'id',
-      'genericList'
-    ],
-    setters: const [
-      'id',
-      'genericList'
-    ]);
+const PageClassMirror = const ClassMirror(name: 'Page', constructors: const {
+  '': const FunctionMirror(name: '', $call: _Page__Constructor)
+}, fields: const {
+  'size': $$Page_fields_size,
+  'total': $$Page_fields_total,
+  'number': $$Page_fields_number,
+  'items': $$Page_fields_items
+}, getters: const [
+  'size',
+  'total',
+  'number',
+  'items'
+], setters: const [
+  'size',
+  'total',
+  'number',
+  'items'
+]);
 
 _Person__Constructor([positionalParams, namedParams]) => new Person();
 
@@ -125,9 +142,8 @@ const PersonClassMirror = const ClassMirror(
 // **************************************************************************
 
 _initMirrors() {
-  initClassMirrors(
-      {SomeObject: SomeObjectClassMirror, Person: PersonClassMirror});
+  initClassMirrors({Page: PageClassMirror, Person: PersonClassMirror});
 
   getClassMirrorFromGenericInstance =
-      (instance) => instance is SomeObject ? SomeObjectClassMirror : null;
+      (instance) => instance is Page ? PageClassMirror : null;
 }